pyproject.toml
[project]
name = "assistingcalibration"
version = "0.1.0"
description = "Python package for assisting calibration tasks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "botorch>=0.16.1",
    "et-xmlfile==2.0.0",
    "gpytorch>=1.14.3",
    "iniconfig==2.3.0",
    "numpy==2.3.5",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pandas==2.3.3",
    "pluggy==1.6.0",
    "pygments==2.19.2",
    "pytest==9.0.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "six==1.17.0",
    "torch>=2.9.1",
    "tzdata==2025.3",
]
